+++ gameInfos/versions.py
@@ -21,7 +21,9 @@ GAME_VERSIONS = {
     1045 : ["a19"],
     1057 : ["a20"],
     1064 : ["a21","a21.1"],
-    1067 : ["a22.2"]
+    1067 : ["a22.2"],
+    1071 : ["a22.3","a22.4"],
+    99999 : ["a22.4"]
 }
 LATEST_GAME_VERSION = next(iter(reversed(GAME_VERSIONS.keys())))
 LATEST_PUBLIC_GAME_VERSION = 1033