+++ main.py
@@ -1,5 +1,5 @@
 # python 3.11.0
-# discord.py 2.3.2
+# discord.py 2.4.0
 # pygame 2.5.2
 import bot
 if __name__ == "__main__":